The Grand Hackathon
The goal of this hackathon is to predict GreatStone’s rating of a mutual fund. In order to help investors decide on which mutual fund to pick for an investment, the task is to build a model that can predict the rating of a mutual fund. The various attributes that define a mutual fund can be used for building the model.
Portal :
https://great-learning.shinyapps.io/AIML_Hackathon_GreatStonePrediction/
Username: kumar.ayush.1989@hotmail.com
Password: 39805